chore: move stuff in better places
