+++ frontend/package.json
@@ -17,7 +17,7 @@
 		"test:unit": "vitest --project unit",
 		"test:nuxt": "vitest --project nuxt"
 	},
-	"packageManager": "pnpm@10.32.0",
+	"packageManager": "pnpm@11.1.1+sha512.d1fdf5f73c617b64fa1a56a81c3c8dfe0e966e33a6010aa256b517ae77be21d93e05affc0de1a83b0e4f29d569f68b446ae8f068cd7247c0bb3df0fb4d7bdf9a",
 	"dependencies": {
 		"@iconify-json/lucide": "^1.2.98",
 		"@iconify-json/simple-icons": "^1.2.74",
