+++ server/src/main/kotlin/org/elaastix/server/core/infrastructure/config/ApiVersioningConfig.kt
@@ -17,7 +17,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package org.elaastix.server.infrastructure.config
+package org.elaastix.server.core.infrastructure.config
 
 import jakarta.servlet.http.HttpServletRequest
 import org.elaastix.server.ElaastixServerApplication
